Fixed miscellaneous issues

+++ src/editor/entity_editor/editor_view.cpp
@@ -41,8 +41,8 @@ void EditorView::update() {
         std::pair<float, float> new_pos;
         auto distance = mouse_.getMouseDistance();
 
-        new_pos.first = initial_mouse_pos_.first - distance.first;
-        new_pos.second = initial_mouse_pos_.second - distance.second;
+        new_pos.first = initial_mouse_pos_.first - static_cast<float>(distance.first);
+        new_pos.second = initial_mouse_pos_.second - static_cast<float>(distance.second);
 
         setView(new_pos, size_);
     }
